move one level up utils folder
